requirements.txt
easyocr==1.6.2
fastapi==0.89.1
numpy==1.21.4
opencv_python_headless==4.5.2.52
Pillow==9.4.0
starlette==0.17.1
